Extract Dialog container

+++ app/src/main/java/com/lukasz/witkowski/training/planner/ui/createTraining/CreateTrainingViewModel.kt
@@ -80,7 +80,7 @@ class CreateTrainingViewModel @Inject constructor(
             exercise = exercise,
             repetitions = repetitions,
             sets = sets,
-            time = timeInMillis.toLong()
+            time = timeInMillis
         )
         addTrainingExercise(trainingExercise)
     }
